Fix cross-entry-point imports in src/link/ws.
